MERN Auth Lesson 15

+++ frontend/src/components/WorkoutDetails.js
@@ -1,14 +1,23 @@
 import useWorkoutsContext from "../hooks/useWorkoutContext";
+import useAuthContext from "../hooks/useAuthContext";
 
 // date fns
 import formatDistanceToNow from "date-fns/formatDistanceToNow";
 
 export default function WorkoutDetails({ workout }) {
   const { dispatch } = useWorkoutsContext();
+  const { user } = useAuthContext();
 
   async function handleClick() {
+    if (!user) {
+      return;
+    }
+
     const response = await fetch("/api/workouts/" + workout._id, {
       method: "DELETE",
+      headers: {
+        Authorization: `Bearer ${user.token}`,
+      },
     });
     const json = await response.json();
